requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM APIs
openai==1.6.0
anthropic==0.7.0
google-generativeai==0.3.0
langchain==0.1.0
langchain-openai==0.0.2
langsmith==0.0.70

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# API Framework
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.0

# Dashboard
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4

# Utilities
requests==2.31.0
aiohttp==3.9.1
tiktoken==0.5.2
tenacity==8.2.3
